useRoomとuseRoomIdを分離
useRoomはこれから404 roomの対策に使う
